--- .streamlit/.streamlit/secrets.toml
@@ -1,1 +1,0 @@
-api_url = "https://immo-eliza-deployment-20bn.onrender.com/predict"
