+++ data/default_values.csv
@@ -1,0 +1,2 @@
+subproperty_type,locality,equipped_kitchen,state_building,epc,fl_terrace,fl_open_fire,fl_swimming_pool,fl_garden,fl_double_glazing,construction_year,latitude,longitude,total_area_sqm,surface_land_sqm,nbr_frontages,nbr_bedrooms,terrace_sqm,primary_energy_consumption_sqm,cadastral_income,garden_sqm,zip_code
+MISSING,MISSING,MISSING,MISSING,MISSING,0,0,0,0,0,1984.4060541310541,50.886624644454066,4.32866000982109,163.67074643572522,1157.0875047764616,2.802928912844503,2.787276026009456,11.577383719998076,1688.7480021044896,1885.9412977998952,115.64028826544673,5144.609341685317
